[wip] design how reading things will work
Threw together a component to try connecting to, this shouldn't change.

Added a read map for what I'd like to do. This will probably change with
how this is made and how other actions are implemented.

Because it gets the one document, It will need to be mapped if I want it
to be a single prop.
I might ditch this and have each doc as it's own prop, but right now I
don't imagine it being very much more difficult.